Rework Plushie's process creation and management
Plushie will now use three processes instead of two. One network one,
one bot one, and the overall process. This includes configuration
options to make it work with Windows and in debug mode automatically.
There are a bunch of other small changes, but for the most part it
should still work the same.